# openapi.yaml
openapi: 3.0.3
info:
  title: Flask CRUD
  version: 1.0.0
...
paths:
  /attack_types:
    post:
      summary: Create an attack type
      tags:
        - AttackTypes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  example: Phishing
      responses:
        "201":
          description: Attack type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackType"
        "400":
          description: Invalid type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List all attack types
      tags:
        - AttackTypes
      responses:
        "200":
          description: List of attack types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttackType"
  /attack_types/{id}:
    get:
      summary: Get an attack type by ID
      tags:
        - AttackTypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Attack type found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackType"
        "404":
          description: Attack type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update an attack type
      tags:
        - AttackTypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  example: Ransomware
      responses:
        "200":
          description: Attack type updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackType"
        "400":
          description: Invalid type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Attack type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete an attack type
      tags:
        - AttackTypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "204":
          description: Attack type deleted
        "404":
          description: Attack type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Attack type cannot be deleted due to foreign key constraint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cyber_threats:
    post:
      summary: Create a cyber threat
      tags:
        - CyberThreats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - country
                - year
                - attack_type
                - target_industry
                - financial_loss
                - affected_users
                - attack_source
                - security_vulnerability
                - defense_mechanism
                - resolution_time
              properties:
                country:
                  type: string
                  example: Brazil
                year:
                  type: integer
                  example: 2023
                attack_type:
                  type: integer
                  example: 1
                target_industry:
                  type: integer
                  example: 2
                financial_loss:
                  type: number
                  format: float
                  example: 12.5
                affected_users:
                  type: integer
                  example: 15000
                attack_source:
                  type: integer
                  example: 3
                security_vulnerability:
                  type: integer
                  example: 4
                defense_mechanism:
                  type: integer
                  example: 1
                resolution_time:
                  type: integer
                  example: 72
      responses:
        "201":
          description: Cyber threat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CyberThreat"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List all cyber threats
      tags:
        - CyberThreats
      responses:
        "200":
          description: List of cyber threats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CyberThreat"
  /cyber_threats/{id}:
    get:
      summary: Get a cyber threat by ID
      tags:
        - CyberThreats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Cyber threat found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CyberThreat"
        "404":
          description: Cyber threat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a cyber threat
      tags:
        - CyberThreats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                country:
                  type: string
                  example: Brazil
                year:
                  type: integer
                  example: 2024
                attack_type:
                  type: integer
                  example: 1
                target_industry:
                  type: integer
                  example: 2
                financial_loss:
                  type: number
                  format: float
                  example: 8.3
                affected_users:
                  type: integer
                  example: 4200
                attack_source:
                  type: integer
                  example: 3
                security_vulnerability:
                  type: integer
                  example: 4
                defense_mechanism:
                  type: integer
                  example: 2
                resolution_time:
                  type: integer
                  example: 48
      responses:
        "200":
          description: Cyber threat updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CyberThreat"
        "404":
          description: Cyber threat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a cyber threat
      tags:
        - CyberThreats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "204":
          description: Cyber threat deleted
        "404":
          description: Cyber threat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    AttackType:
      type: object
      properties:
        id:
          type: integer
          example: 1
        type:
          type: string
          example: Phishing
    CyberThreat:
      type: object
      properties:
        Id:
          type: integer
        Country:
          type: string
        Year:
          type: integer
        Financial Loss (in Million $):
          type: number
        Number of Affected Users:
          type: integer
        Incident Resolution Time (in Hours):
          type: integer
        Attack Type:
          type: object
          properties:
            Id:
              type: integer
            Type:
              type: string
        Target Industry:
          type: object
          properties:
            Id:
              type: integer
            Industry:
              type: string
        Attack Source:
          type: object
          properties:
            Id:
              type: integer
            Source:
              type: string
        Security Vulnerability Type:
          type: object
          properties:
            Id:
              type: integer
            Vulnerability:
              type: string
        Defense Mechanism Used:
          type: object
          properties:
            Id:
              type: integer
            Mechanism:
              type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Not found